+++ test/vanilla/legacy/Expected/AcceptanceTests/Validation/validation/aio/operations/_auto_rest_validation_test_operations.py
@@ -57,7 +57,7 @@ class AutoRestValidationTestOperationsMixin:
         _headers = kwargs.pop("headers", {}) or {}
         _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
-        api_version = kwargs.pop("api_version", _params.pop("apiVersion", "1.0.0"))  # type: str
+        api_version = kwargs.pop("api_version", _params.pop("apiVersion", self._config.api_version))  # type: str
         cls = kwargs.pop("cls", None)  # type: ClsType[_models.Product]
 
         request = build_validation_of_method_parameters_request(
@@ -175,7 +175,7 @@ class AutoRestValidationTestOperationsMixin:
         _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
         _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
-        api_version = kwargs.pop("api_version", _params.pop("apiVersion", "1.0.0"))  # type: str
+        api_version = kwargs.pop("api_version", _params.pop("apiVersion", self._config.api_version))  # type: str
         content_type = kwargs.pop("content_type", _headers.pop("Content-Type", None))  # type: Optional[str]
         cls = kwargs.pop("cls", None)  # type: ClsType[_models.Product]